requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
mysql-connector-python @ http://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.0.4.zip
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
